+++ lib/src/extension_set.dart
@@ -7,6 +7,15 @@ class ExtensionSet {
   static ExtensionSet commonMark = new ExtensionSet._(
       [const FencedCodeBlockSyntax()], [new InlineHtmlSyntax()]);
 
+  static ExtensionSet gitHub = new ExtensionSet._([
+    const FencedCodeBlockSyntax(),
+    const HeaderWithIdSyntax(),
+    const SetextHeaderWithIdSyntax(),
+    const TableSyntax()
+  ], [
+    new InlineHtmlSyntax()
+  ]);
+
   final List<BlockSyntax> blockSyntaxes;
   final List<InlineSyntax> inlineSyntaxes;